[consensus] may2023Height is defined otherwhise code won't compile
The removed `if` check was problematic because on regtest
`consensusparams.may2023Height` is set to 0 as all other activation
heights. Hence checking if the variables exists via:

    if (consensusparams.may2023Height)

always turns to be false.

The check didn't make sense to me anyway since if `may2023Height`
were not defined then the source code would not compile.