Remove object to track the exact height when May 2023 upgrade activated
This code was used to keep track of May 2023 activation height so that
the node would not have go trough the MTP activation logic every time
it checks for upgrade activation. This was particularly useful for the
code path where we are certain that the upgrade is activated but we still
need activation height

Now that we convert May 2023 activation to be height based and also we
chose not to use this cache technique for the May 2024 upgrade, we don't
need the code anymore.