[qa] temp fix for feature and wallet p2sh32
there's to be an off by one error somewhere in the mempool admission
code when evaluating the activation state of may 2023 (upgrade9) protocol
upgrade.

usually an upgrade is activated on the last block before the new
rule will be enforced at the blok level. That means that for mempool
transactions at that height the validity rules are already changed.

To pass those 2 tests I had to set activation height a block earlier
than necesary, this implies that mempool validation is applying
the new rules after the first block after activation is mined.

This is not a big deal thou, cause it owuld not bring any fork
in the netwrok, just the rejection of txs that would be valid for
the upgrade9 rules for 10 minutes or so.

Will try to fix late, once we have anythign in place.