[qa] Fix the last issue for feature_abla
Now this assert works after we call `getblocktemplate` set to
prodxeuce a block half of the max block size (according to abla)

assert_greater_than_or_equal(mining_limit, mi['currentblocksize'])

I needed to put a sleep before gbtl call to make it work thou.
Without it, it was lihke the getblocktemplate cmd was not executed.