ABLA: Future proof code against potential for bug in CDiskBlockIndex serialization
Summary
---

The code for `CDiskBlockIndex` related to ABLA was potentially clearing
the ABLA state for the `CDiskBlockIndex` even in the *write* case.

This MR restricts the clearing of the ABLA state to only happen in the
*reading* case, and never *write*.

For this potential bug to be triggered, the following condition
would need to be met: Serializer has its `GetVersion()` set below
`CLIENT_VERSION_ABLA_BLOCKINDEX_DATA`.

This condition is never met in the current codebase.

Moreover, in the current codebase, the serialization happens to a
*temporary copy* of the real `CBlockIndex` being serialized, casted to
`CDiskBlockIndex`, so even if this potential bug were triggered, the
ABLA state being cleared here would be harmless, since it would only be
cleared for this ephemeral copy.

However, in order to prevent potential bugs should the above assumptions
change, I am submitting this nit fix MR.

Test Plan
---

This MR does not modify behavior in any way. However:

- `make && make check`